Added documentation for the command listener and renamed it interaction listener because discord listeners have a limit and affect performance. Also moved it up a folder to reflect that change
